ZOMG MEGA Commit
So I added the Korg Volca Bass, which in turn lead me to refactor my
code as it was too smart for it's own good. Which @csexton calls
Coding regret.

After a big refactor to make things much simpler, we have some
code repitition, but that's ok we can remove that later.

We also included the ability to load in via a yaml file which
instrument is connected to which midi interface.

Then we have a demo to wrap them gem up in a DRB server so that we
can controll all the Volcas via Sonic Pi.